Update control_structures.md to include clarity on case/2 parameters
Added variable assignment before case/2 examples. As a first time Elixir learner, I was confused by the use of the tuple and introduction to _, as it was unclear to me that "case {value} do"
was using the example value (i.e. {:ok, "Hello World"}, :even) to pattern match. I added a variable assignment before as this is generally as I see case being useful